+++ examples/data-sources/iosxe_ip_host/data-source.tf
@@ -1,2 +1,0 @@
-data "iosxe_ip_host" "example" {
-}