+++ examples/resources/iosxe_ip_host/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxe_ip_host.example ""