+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.7.1 (unreleased)
 
 - Change type of `transport_input` attribute of `iosxe_line` resource and data source from `String` to `List of String`
-- Add `ip_host_lists`, `ip_host_vrfs`, and `diagnostic_event_log_size` attributes to `iosxe_system` resource and data source
+- Add `ip_hosts`, `ip_hosts_vrfs`, `name`, `ips`, `vrf`, `hosts`, `subscriber_templating`, `call_home_contact_email`, `call_home_cisco_tac_1_profile_active`, `call_home_cisco_tac_1_destination_transport_method`, `ip_ftp_passive`, `tftp_source_interface_gigabit_ethernet`, `tftp_source_interface_loopback`, `version`, and `multilink_ppp_bundle_name` to `iosxe_system` resource and data source
 - Add `ip_ssh_version_legacy`, `ip_ssh_time_out` and `ip_ssh_authentication_retries` attributes to `iosxe_system` resource and data source
 
 ## 0.7.0